% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aegypticdmx-package.R
\docType{package}
\name{aegypticdmx-package}
\alias{aegypticdmx}
\alias{aegypticdmx-package}
\title{aegypticdmx: The package provides data on the presence of Aedes aegypti in the metropolitan area of Mexico City}
\description{
The package provides data on the presence of Aedes aegypti in the metropolitan area of Mexico City and machine learning algorithms to predict the presence of Aedes aegypti.
}
\author{
\strong{Maintainer}: Felipe Antonio Dzul Manzanilla \email{felipe.dzul.m@gmail.com} (\href{https://orcid.org/0000-0002-2085-9141}{ORCID})

Other contributors:
\itemize{
  \item Fabian Correa-Morales \email{fabiancorrea@msn.com} (\href{https://orcid.org/0000-0002-6193-1242}{ORCID}) [contributor]
}

}
\keyword{internal}
