% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ua_cdmx}
\alias{ua_cdmx}
\title{ua_cdmx}
\format{
A sf object with 341 rows and 30 variables:
\describe{
  \item{id}{id of cdmx (383.}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://dataforgood.facebook.com/dfg/tools/globalurbanareas}
}
\usage{
ua_cdmx
}
\description{
global urban area of cdmx
}
\details{
The urban area of the metropolitan area of Mexico City was the global urban areas of Facebook meta
}
\keyword{datasets}
